# public/swagger.json
{
	"components": {
...
		"schemas": {
			"RegisterBodyAuthType": {
				"properties": {
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"lastname": {
						"type": "string"
					},
					"firstname": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"LoginBodyAuthType": {
				"properties": {
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"Partial_ICourse_": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"miniature": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"Types.ObjectId": {
				"type": "string"
			},
			"Partial___id-Types.ObjectId__": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/Types.ObjectId"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"ObjCourseType": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Partial_ICourse_"
					},
					{
						"$ref": "#/components/schemas/Partial___id-Types.ObjectId__"
					}
				]
			},
			"CreateBodyCourseType": {
				"properties": {
					"score": {
						"type": "number",
						"format": "double"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"url": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"UpdateBodyCourseType": {
				"properties": {
					"score": {
						"type": "number",
						"format": "double"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"url": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"Partial_IMenu_": {
				"properties": {
					"title": {
						"type": "string"
					},
					"path": {
						"type": "string"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"active": {
						"type": "boolean"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"ObjMenuType": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Partial_IMenu_"
					},
					{
						"$ref": "#/components/schemas/Partial___id-Types.ObjectId__"
					}
				]
			},
			"CreateBodyMenuType": {
				"properties": {
					"order": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"active": {
						"type": "boolean"
					},
					"path": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"UpdateBodyMenuType": {
				"properties": {
					"order": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"active": {
						"type": "boolean"
					},
					"path": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "back-personal-web-mern",
		"version": "1.0.0",
		"description": "- [Get the Project](#get-the-project)\r - [Project Dependencies](#project-dependencies)\r - [Project Scripts](#project-scripts)\r - [Environment Variables](#environment-variables)\r - [Authors ](#authors )\r - [License](#license)",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "S.U."
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/auth/register": {
			"post": {
				"operationId": "RegisterUser",
				"responses": {
					"200": {
						"description": "Message informing if create was correct",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Endpoint to create new user in database",
				"tags": [
					"AuthController"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "new user parameters (required)",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterBodyAuthType",
								"description": "new user parameters (required)"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "LoginUser",
				"responses": {
					"200": {
						"description": "Message with jwt access token and jwt refresh token",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"refreshToken": {
											"type": "string"
										},
										"accessToken": {
											"type": "string"
										}
									},
									"required": [
										"refreshToken",
										"accessToken"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Endpoint to auth with login a user in database",
				"tags": [
					"AuthController"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "credentials of the user to log in (required)",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginBodyAuthType",
								"description": "credentials of the user to log in (required)"
							}
						}
					}
				}
			}
		},
		"/auth/refresh": {
			"post": {
				"operationId": "RefreshAccessToken",
				"responses": {
					"200": {
						"description": "Message with a new jwt access token",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"accessToken": {
											"type": "string"
										}
									},
									"required": [
										"accessToken"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Endpoint to refresh the token of a previously logged in user",
				"tags": [
					"AuthController"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Valid Refresh token",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "string",
								"description": "Valid Refresh token"
							}
						}
					}
				}
			}
		},
		"/course": {
			"get": {
				"operationId": "GetCourses",
				"responses": {
					"200": {
						"description": "Course data according to the filters applied",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"pages": {
											"type": "number",
											"format": "double"
										},
										"page": {
											"type": "number",
											"format": "double"
										},
										"limit": {
											"type": "number",
											"format": "double"
										},
										"total": {
											"type": "number",
											"format": "double"
										},
										"docs": {
											"items": {
												"$ref": "#/components/schemas/ObjCourseType"
											},
											"type": "array"
										}
									},
									"required": [
										"pages",
										"page",
										"limit",
										"total",
										"docs"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Endpoint to obtain the data of courses from database",
				"tags": [
					"CourseController"
				],
				"security": [],
				"parameters": [
					{
						"description": "maximum number of courses to retrieve per page (optional)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "page number to return (optional)",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateCourse",
				"responses": {
					"200": {
						"description": "Data of the new course created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ObjCourseType"
								}
							}
						}
					}
				},
				"description": "Endpoint to create courses in datatbase",
				"tags": [
					"CourseController"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "new course parameters (required)",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBodyCourseType",
								"description": "new course parameters (required)"
							}
						}
					}
				}
			}
		},
		"/course/{id}": {
			"patch": {
				"operationId": "UpdateCourse",
				"responses": {
					"200": {
						"description": "Message informing if update was correct",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Endpoint to partially update course information from the database",
				"tags": [
					"CourseController"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "identifier of the course to modify (required)",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "parameters of the course to modify (Optional)",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateBodyCourseType",
								"description": "parameters of the course to modify (Optional)"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCourse",
				"responses": {
					"200": {
						"description": "Message informing if delete was correct",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Endpoint to delete course information from the database",
				"tags": [
					"CourseController"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "identifier of the course to delete (required)",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/menu": {
			"get": {
				"operationId": "GetMenus",
				"responses": {
					"200": {
						"description": "Information of all menus with or without filters",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ObjMenuType"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Endpoint to obtain the data of menus from database",
				"tags": [
					"MenuController"
				],
				"security": [],
				"parameters": [
					{
						"description": "active menus filter you want to get (true or false) (Optional)",
						"in": "query",
						"name": "active",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateMenu",
				"responses": {
					"200": {
						"description": "Data of the new menu",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ObjMenuType"
								}
							}
						}
					}
				},
				"description": "Endpoint to create menu from database",
				"tags": [
					"MenuController"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "parameter to new menu (required)",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBodyMenuType",
								"description": "parameter to new menu (required)"
							}
						}
					}
				}
			}
		},
		"/menu/{id}": {
			"patch": {
				"operationId": "UpdateMenu",
				"responses": {
					"200": {
						"description": "Message informing if update was correct",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Endpoint to partially update menu information from the database",
				"tags": [
					"MenuController"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "identifier of the menu to modify (required)",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "parameters of the menu to modify (Optional)",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateBodyMenuType",
								"description": "parameters of the menu to modify (Optional)"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMenu",
				"responses": {
					"200": {
						"description": "Message informing if delete was correct",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Endpoint to delete menu information from the database",
				"tags": [
					"MenuController"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "identifier of the menu to delete (required)",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}